backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
python-multipart==0.0.20
scikit-learn==1.6.1
spacy==3.8.4
pandas==2.2.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
joblib==1.4.2
